requirements.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
bandit==1.7.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.44.0
django-filter==2.4.0
django-heroku==0.3.1
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.1.1
oauthlib==3.1.1
pbr==5.6.0
psycopg2==2.9.3
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.6
python-dateutil==1.5
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
urllib3==1.26.6
whitenoise==5.2.0
wrapt==1.12.1
